Moved testing from main to testailu.c also created testailu.h
