+++ Debug/aaro_larange2.log
@@ -1,5 +1,4 @@
 ﻿  main.c
-  polynomit.c
+  testailu.c
   Generating Code...
-  LINK : C:\Users\aaroi\Desktop\larange_lukija_uusi\aaro_larange2\Debug\aaro_larange2.exe not found or not built by the last incremental link; performing full link
   aaro_larange2.vcxproj -> C:\Users\aaroi\Desktop\larange_lukija_uusi\aaro_larange2\Debug\aaro_larange2.exe
