+++ aaro_larange2.vcxproj
@@ -141,9 +141,11 @@
   <ItemGroup>
     <ClCompile Include="main.c" />
     <ClCompile Include="polynomit.c" />
+    <ClCompile Include="testailu.c" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="polynomit.h" />
+    <ClInclude Include="testailu.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
