+++ testailu.h
@@ -1,0 +1,7 @@
+#ifndef	TESTAILU_H
+#define	TESTAILU_H
+
+void kokeilu1(void);
+
+
+#endif //TESTAILU_H
